move frontend-service to its own folder in services
